# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EarningsDesk API",
    "description": "Read-only earnings + options data for the EarningsDesk dashboard.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/calendar": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "Get Calendar",
        "operationId": "get_calendar_api_calendar_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 14,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ticker/{ticker}": {
      "get": {
        "tags": [
          "ticker"
        ],
        "summary": "Get Ticker",
        "operationId": "get_ticker_api_ticker__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "with_prices",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "fetch the price chart series",
              "default": true,
              "title": "With Prices"
            },
            "description": "fetch the price chart series"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TickerPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalendarEntry": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "report_date": {
            "type": "string",
            "format": "date",
            "title": "Report Date"
          },
          "session": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session"
          },
          "days_until": {
            "type": "integer",
            "title": "Days Until"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "implied_move": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implied Move"
          },
          "hist_avg_move": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hist Avg Move"
          },
          "edge_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edge Score"
          },
          "spot": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spot",
            "description": "current price, not point-in-time at report_date"
          },
          "atm_open_interest": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atm Open Interest",
            "description": "contracts at the at-the-money strike, latest snapshot; low = wide spreads, hard to fill"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "report_date",
          "days_until"
        ],
        "title": "CalendarEntry"
      },
      "CalendarPage": {
        "properties": {
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of"
          },
          "window_days": {
            "type": "integer",
            "title": "Window Days"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/CalendarEntry"
            },
            "type": "array",
            "title": "Entries",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "window_days"
        ],
        "title": "CalendarPage"
      },
...
      "EarningsHistoryRow": {
        "properties": {
          "report_date": {
            "type": "string",
            "format": "date",
            "title": "Report Date"
          },
          "session": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session",
            "description": "\"BMO\" | \"AMC\" | null when unknown"
          },
          "eps_estimate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eps Estimate"
          },
          "eps_actual": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eps Actual"
          },
          "eps_surprise": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eps Surprise",
            "description": "percent, e.g. 5.54 == +5.54%"
          },
          "implied_move": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implied Move",
            "description": "fraction, e.g. 0.078 == \u00b17.8%"
          },
          "realized_move": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Realized Move",
            "description": "signed close-to-close fraction"
          },
          "beat_implied": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beat Implied"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict",
            "description": "RICH | CHEAP | FAIR"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction",
            "description": "BULLISH | BEARISH | NEUTRAL"
          },
          "correct_direction": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correct Direction"
          },
          "gap_open_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gap Open Pct"
          },
          "gap_filled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gap Filled"
          },
          "vol_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vol Ratio",
            "description": "earnings-day volume / 20d average"
          },
          "pnl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pnl",
            "description": "long-straddle proxy return"
          }
        },
        "type": "object",
        "required": [
          "report_date"
        ],
        "title": "EarningsHistoryRow",
        "description": "One past (or upcoming) earnings event, everything we know about it."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryStats": {
        "properties": {
          "n_events": {
            "type": "integer",
            "title": "N Events"
          },
          "hit_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hit Rate"
          },
          "avg_pnl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Pnl"
          },
          "avg_implied_move": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Implied Move"
          },
          "avg_realized_move": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Realized Move"
          },
          "overpricing_bias": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overpricing Bias",
            "description": "positive = market consistently overestimates the move"
          },
          "avg_gap_open_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Gap Open Pct"
          },
          "gap_fill_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gap Fill Rate"
          },
          "avg_vol_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Vol Ratio"
          }
        },
        "type": "object",
        "required": [
          "n_events"
        ],
        "title": "HistoryStats",
        "description": "Aggregates over the earnings history. Mirrors core.llm_context.TickerContext.\n\n`hit_rate` is null below 4 scored events on purpose \u2014 a single correct call\nreads as 100% accuracy and that number would get screenshotted. See\nMIN_EVENTS_FOR_HIT_RATE."
      },
...
      "OptionsPanel": {
        "properties": {
          "implied_move": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implied Move"
          },
          "hist_avg_move": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hist Avg Move"
          },
          "richness": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Richness",
            "description": "implied / historical - 1; +0.30 means 30% richer than typical"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "edge_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edge Score",
            "description": "0-10"
          },
          "put_call_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Put Call Ratio"
          },
          "atm_open_interest": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atm Open Interest"
          },
          "atm_strike": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atm Strike"
          },
          "atm_expiry": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atm Expiry"
          },
          "call_volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Call Volume",
            "description": "today's total call contracts traded"
          },
          "put_volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Put Volume",
            "description": "today's total put contracts traded"
          },
          "atm_volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atm Volume",
            "description": "today's call+put volume at the ATM strike, vs. atm_open_interest"
          },
          "oi_by_strike": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/StrikeOpenInterest"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oi By Strike",
            "description": "a window of strikes around the ATM, for a liquidity chart"
          },
          "skew_put_iv": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skew Put Iv",
            "description": "IV of the ~25-delta put"
          },
          "skew_call_iv": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skew Call Iv",
            "description": "IV of the ~25-delta call"
          },
          "skew_risk_reversal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skew Risk Reversal",
            "description": "put IV minus call IV at ~25 delta. Positive = downside protection is the expensive side, the ordinary state for equities, so only an unusual level is informative. Describes option pricing, not a directional forecast."
          },
          "iv_front": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iv Front"
          },
          "iv_back": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iv Back"
          },
          "iv_inverted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iv Inverted",
            "description": "front IV above back IV \u2014 the earnings premium"
          }
        },
        "type": "object",
        "title": "OptionsPanel",
        "description": "What the options market is pricing right now (as of the snapshot)."
      },
...
      "TickerPage": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "is_tracked": {
            "type": "boolean",
            "title": "Is Tracked",
            "description": "False = outside the universe; no signal track record exists",
            "default": true
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "snapshot_age_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Age Hours"
          },
          "is_stale": {
            "type": "boolean",
            "title": "Is Stale",
            "default": false
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "company_domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Domain",
            "description": "for a logo lookup, e.g. 'walmart.com'"
          },
          "spot": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spot"
          },
          "next_report_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Report Date"
          },
          "next_report_session": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Report Session"
          },
          "days_until_report": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Until Report"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction",
            "description": "BULLISH | BEARISH | NEUTRAL \u2014 the validated options-flow + sentiment lean for the *next* report, refined daily starting a few days out. Absent (not NEUTRAL) until that window opens; describes what the market's flow suggests, not a recommendation."
          },
          "direction_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction Confidence",
            "description": "0-1"
          },
          "direction_as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction As Of",
            "description": "the date this read was last refreshed, for an age/freshness label"
          },
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptionsPanel"
              },
              {
                "type": "null"
              }
            ]
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/EarningsHistoryRow"
            },
            "type": "array",
            "title": "History",
            "default": []
          },
          "stats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HistoryStats"
              },
              {
                "type": "null"
              }
            ]
          },
          "prices": {
            "items": {
              "$ref": "#/components/schemas/PricePoint"
            },
            "type": "array",
            "title": "Prices",
            "default": []
          },
          "news": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/NewsItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "News",
            "description": "null = fetch failed; [] = genuinely no news"
          },
          "news_sentiment": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "News Sentiment",
            "description": "-1..1"
          },
          "analyst_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst Score",
            "description": "-1..1"
          },
          "analyst_rating_raw": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst Rating Raw",
            "description": "Seeking Alpha 1-5"
          },
          "past_moves": {
            "items": {
              "$ref": "#/components/schemas/PastMove"
            },
            "type": "array",
            "title": "Past Moves",
            "description": "Realized move per past print, oldest first. Empty until the nightly build populates it \u2014 a normal not-yet state, not an error.",
            "default": []
          },
          "peers": {
            "items": {
              "$ref": "#/components/schemas/PeerEarnings"
            },
            "type": "array",
            "title": "Peers",
            "description": "Related companies (curated similar-product peers, else same-sector tracked names) with their most recent completed print's before/after prices. Empty when no peer has usable recent-earnings data yet.",
            "default": []
          },
          "fundamentals": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Fundamentals"
              },
              {
                "type": "null"
              }
            ],
            "description": "null = the lookup failed or the symbol is unknown"
          },
          "analyst_ratings": {
            "items": {
              "$ref": "#/components/schemas/AnalystRatingRow"
            },
            "type": "array",
            "title": "Analyst Ratings",
            "description": "newest first; empty = no coverage or the lookup failed",
            "default": []
          },
          "ai_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AiSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "ticker"
        ],
        "title": "TickerPage"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}